refactor: decouple page labels from PageSelector
